Develop

+++ Projects/ProjectKB/Config/SummaryInfo.cs
@@ -50,13 +50,13 @@ namespace ProjectKB.Config
         /// 良品率
         /// </summary>
         [JsonIgnore]
-        public double GoodProductRate { get => ActualProduction > 0 ? (double)GoodProductCount / (double)ActualProduction : 0; }
+        public double GoodProductRate { get => ActualProduction > 0 ? GoodProductCount / (double)ActualProduction : 0; }
 
         /// <summary>
         /// 不良率
         /// </summary>
         [JsonIgnore]
-        public double DefectiveProductRate { get => ActualProduction > 0 ? (double)DefectiveProductCount / (double)ActualProduction : 0; }
+        public double DefectiveProductRate { get => ActualProduction > 0 ? DefectiveProductCount / (double)ActualProduction : 0; }
